+++ ttnn/cpp/ttnn/operations/experimental/ccl/llama_reduce_scatter/device/llama_reduce_scatter_program_factory.cpp
@@ -656,11 +656,11 @@ LlamaReduceScatterDeviceOperation::LlamaReduceScatterAdd::create_at_program_proc
         num_devices,
         input_page_size,
         output_cores_per_device,
-        packet_start_worker_core.at(0).x,
-        packet_start_worker_core.at(0).y,
-        packet_end_worker_core.at(0).x,
-        packet_end_worker_core.at(0).y,
-        sender_cores.size(),
+        static_cast<uint32_t>(packet_start_worker_core.at(0).x),
+        static_cast<uint32_t>(packet_start_worker_core.at(0).y),
+        static_cast<uint32_t>(packet_end_worker_core.at(0).x),
+        static_cast<uint32_t>(packet_end_worker_core.at(0).y),
+        static_cast<uint32_t>(sender_cores.size()),
         total_num_read_txns,
         signaler.has_value()};
 
@@ -705,9 +705,9 @@ LlamaReduceScatterDeviceOperation::LlamaReduceScatterAdd::create_at_program_proc
         num_devices,
         input_page_size,
         output_cores_per_device,
-        packet_receiver_worker_core.x,
-        packet_receiver_worker_core.y,
-        num_packet_worker_cores,
+        static_cast<uint32_t>(packet_receiver_worker_core.x),
+        static_cast<uint32_t>(packet_receiver_worker_core.y),
+        static_cast<uint32_t>(num_packet_worker_cores),
         topology == tt::tt_fabric::Topology::Ring ? 1u : 0u};
 
     auto writer_defines = reader_defines;
@@ -745,7 +745,7 @@ LlamaReduceScatterDeviceOperation::LlamaReduceScatterAdd::create_at_program_proc
     uint32_t local_page = 0;
 
     std::vector<uint32_t> reader_runtime_args = {
-        cross_device_semaphore->address(), local_semaphore, false, false, 0, false, 0, 0, 0};
+        static_cast<uint32_t>(cross_device_semaphore->address()), local_semaphore, false, false, 0, false, 0, 0, 0};
     uint32_t is_reader_sender_core_idx = 2;
     uint32_t is_reader_worker_core_idx = 3;
     uint32_t is_linear_input_packet_start_idx = 4;
@@ -772,7 +772,7 @@ LlamaReduceScatterDeviceOperation::LlamaReduceScatterAdd::create_at_program_proc
 
     for (auto core : all_cores) {
         std::vector<uint32_t> writer_runtime_args = {
-            cross_device_semaphore->address(), local_semaphore, false, false, 0, 0, 0, 0};
+            static_cast<uint32_t>(cross_device_semaphore->address()), local_semaphore, false, false, 0, 0, 0, 0};
 
         uint32_t num_shards_to_read_per_worker = schedule[sender_core_idx].size();
